requirements.txt
cairocffi==0.8.0
CairoSVG==2.0.3
cffi==1.10.0
cssselect==1.0.1
cssselect2==0.2.0
Django==1.8
django-haystack==2.6.1
django-taggit==0.22.1
html5lib==0.999999999
jieba==0.39
lxml==3.8.0
Markdown==2.6.9
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
pycparser==2.18
Pyphen==0.9.4
pytz==2017.2
six==1.10.0
tinycss==0.4
tinycss2==0.6.0
WeasyPrint==0.40
webencodings==0.5.1
Whoosh==2.7.4
